Read in Tracon 2024 data and refactor KompassiProgramItem
