+++ client/src/locales/en.ts
@@ -381,7 +381,7 @@ export const translationEN = {
     experiencePoint: "Experience Point",
     other: "Other program",
     roundtableDiscussion: "Roundtable discussion",
-    fleaMarket: "Flea market time",
+    fleaMarket: "Flea market",
   },
   programTypeSingular: {
     all: "program item",
@@ -436,7 +436,7 @@ export const translationEN = {
     experiencePoint: "games",
     other: "program items",
     roundtableDiscussion: "roundtable discussions",
-    fleaMarket: "flea market time",
+    fleaMarket: "flea market times",
   },
   programTypeInessive: {
     all: "program item",
